+++ changelog.md
@@ -1,3 +1,6 @@
 ### 2021-02-21
 - added orthogonal initializers for conv and fc layers
 - added residual conv blocks with orthogonal initializers
+
+### 2021-03-03
+- added basic residual block and bottleneck block for conv ops